refactor(LottoMachine) : 로또 수동생성 로직 이동
